Removed old pyXLIGHT interface
