+++ pyxlight/__init__.py
@@ -1,13 +1,8 @@
 __version__ = "2.0.0"
 
-try:
-    from .pyXLIGHT_solver import PYXLIGHT
-except ImportError:
-    pass
+from .pyXLIGHT import PYXLIGHT
 
 try:
     from .postprocess import AnimateAirfoilOpt
 except ImportError:
     pass
-
-from .pyXLIGHT import xfoilAnalysis
